Brings the standalone Master Data Configuration and BOM/Recipe/Costing
Laravel modules into this app as the first step toward one unified
15-module Tan90 ERP, with a single login page listing every module's
roles. GRN's own auth (enum-based role column, EnsureRole middleware,
navigation/notifications) is left untouched by design - the two modules
run their own independent tan90_roles-based RBAC side by side, since a
full identity merge is a separate, deliberately deferred piece of work.

Also fixes several bugs surfaced by actually running this for the first
time against real MySQL: two MySQL identifier-length overflows in the
copied migrations, a pre-existing gate_entry_id/gate_entries FK type
mismatch in this app's own schema, users.role being NOT NULL with no
default (blocking every Tan90-only seeded user), a missing named
logout route the modules' shared layout assumed existed, and three
Master Data roles (QC/Finance/Procurement) that had no actual login
account behind them.

+++ database/migrations/2026_07_11_000003_add_type_and_nullable_gate_to_validation_issues_table.php
@@ -37,12 +37,14 @@ return new class extends Migration
             $table->dropColumn('gate_entry_id');
         });
 
-        // gate_entries.id is a plain signed `int` (not bigint/unsigned) in
-        // this schema, so the FK column has to match that exactly or MySQL
-        // rejects it (errno 150) — can't use foreignId()'s
-        // unsignedBigInteger here.
+        // gate_entries.id is Laravel's default bigint unsigned (from
+        // $table->id() in its own create migration), so the FK column has
+        // to match that exactly or MySQL rejects it (errno 150) - a plain
+        // signed integer() here was errno-150-incompatible; only went
+        // unnoticed before because it had never been run against a real
+        // MySQL engine (SQLite doesn't enforce column types this strictly).
         Schema::table('validation_issues', function (Blueprint $table) {
-            $table->integer('gate_entry_id')->nullable()->after('id');
+            $table->unsignedBigInteger('gate_entry_id')->nullable()->after('id');
             $table->foreign('gate_entry_id')->references('id')->on('gate_entries')->cascadeOnDelete();
         });
     }
